nested product map

+++ internal/entity/product.go
@@ -4,6 +4,6 @@ type Product struct {
 	Name        string
 	Description string
 	Image       string
-	Weight      float64
+	Weight      int
 	Price       float64
 }
